chore: rename owners data table
