+++ apps/app/app/(wallet)/[address]/owners/(components)/columns.tsx
@@ -18,8 +18,8 @@
 import { Avatar, Checkbox } from "@lightdotso/ui";
 import type { ColumnDef } from "@tanstack/react-table";
 import { z } from "zod";
-import { DataTableColumnHeader } from "@/app/(wallet)/[address]/owners/(components)/data-table-column-header";
-import { DataTableRowActions } from "@/app/(wallet)/[address]/owners/(components)/data-table-row-actions";
+import { DataTableColumnHeader } from "@/app/(wallet)/[address]/owners/(components)/data-table/data-table-column-header";
+import { DataTableRowActions } from "@/app/(wallet)/[address]/owners/(components)/data-table/data-table-row-actions";
 import { PlaceholderOrb } from "@/components/lightdotso/placeholder-orb";
 
 // -----------------------------------------------------------------------------
