+++ apps/app/app/(wallet)/[address]/owners/(components)/data-table/data-table-toolbar.tsx
@@ -18,8 +18,8 @@
 import { Button, Input } from "@lightdotso/ui";
 import { Cross2Icon } from "@radix-ui/react-icons";
 import type { Table } from "@tanstack/react-table";
-import { DataTableFacetedFilter } from "@/app/(wallet)/[address]/owners/(components)/data-table-faceted-filter";
-import { DataTableViewOptions } from "@/app/(wallet)/[address]/owners/(components)/data-table-view-options";
+import { DataTableFacetedFilter } from "@/app/(wallet)/[address]/owners/(components)/data-table/data-table-faceted-filter";
+import { DataTableViewOptions } from "@/app/(wallet)/[address]/owners/(components)/data-table/data-table-view-options";
 import { MAX_WEIGHT } from "@/const/configuration";
 
 // -----------------------------------------------------------------------------
