+++ apps/app/app/(wallet)/[address]/owners/(components)/owners-data-table.tsx
@@ -20,7 +20,7 @@ import { useSuspenseQuery, useQueryClient } from "@tanstack/react-query";
 import type { FC } from "react";
 import type { Address } from "viem";
 import { columns } from "@/app/(wallet)/[address]/owners/(components)/columns";
-import { DataTable } from "@/app/(wallet)/[address]/owners/(components)/data-table";
+import { DataTable } from "@/app/(wallet)/[address]/owners/(components)/data-table/data-table";
 import type { ConfigurationData } from "@/data";
 import { queries } from "@/queries";
